% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbeta.R
\name{mlbeta}
\alias{mlbeta}
\title{Maximum Likelihood Estimation for Beta Distribution Parameters}
\usage{
mlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{A numeric vector representing the sample data, assumed to be drawn from a beta distribution.}

\item{alpha}{A numeric vector of candidate values for the \code{alpha} parameter of the beta distribution.}

\item{beta}{A numeric vector of candidate values for the \code{beta} parameter of the beta distribution.}

\item{...}{Additional arguments passed to the contour plot.}
}
\value{
A list with the following components:
\describe{
  \item{\code{x}}{The original sample data.}
  \item{\code{coord}}{The coordinates of the MLEs in the \code{alpha} and \code{beta} grids.}
  \item{\code{maxl}}{The maximum likelihood value.}
  \item{\code{maxalpha}}{The MLE for the \code{alpha} parameter.}
  \item{\code{maxbeta}}{The MLE for the \code{beta} parameter.}
}
}
\description{
Computes the maximum likelihood estimates (MLEs) for the parameters of a
beta distribution given a sample and grids of possible parameter values
for \code{alpha} and \code{beta}. This function uses a grid search over
the specified \code{alpha} and \code{beta} values and returns the MLEs
along with their likelihood values.
}
\details{
This function performs a grid search over specified values of \code{alpha}
and \code{beta} to find the maximum likelihood estimates of these parameters.
It computes the log-likelihood values for each combination of \code{alpha} and
\code{beta}, identifies the values that maximize the likelihood, and returns them
along with the maximum likelihood.
}
\examples{
# Sample data generated from a beta distribution
set.seed(123)
x <- rbeta(30, shape1 = 3, shape2 = 4)

# Define grids for alpha and beta
alpha <- seq(2.5, 3.5, length.out = 100)
beta <- seq(3.5, 4.5, length.out = 100)

# Run the mlbeta function
result <- mlbeta(x, alpha, beta)
print(result)

}
