% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval Estimation}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{An integer. The number of bootstrap iterations to perform. Default is 10000.}

\item{x}{A numeric vector. The input data from which to bootstrap.}

\item{fun}{A character string. The name of the statistic function to be applied (e.g., `"mean"`, `"median"`). Default is `"mean"`.}

\item{alpha}{A numeric value. The significance level for the confidence interval. Default is 0.05.}

\item{cx}{A numeric value. Controls the size of the text in the plot. Default is 1.5.}

\item{...}{Additional graphical parameters to be passed to the `hist` function.}
}
\value{
A list with the following elements:
\describe{
  \item{ci}{A numeric vector. The lower and upper bounds of the bootstrap confidence interval.}
  \item{fun}{A character string. The statistic function used.}
  \item{x}{The original data used for bootstrapping.}
}
}
\description{
Generates a bootstrap confidence interval for a specified statistic (default is the mean)
by resampling the input data `x` a specified number of times. A histogram of the bootstrap
sample statistics is also generated.
}
\details{
The function resamples the input data `x` with replacement for a specified number
of iterations (`iter`). The specified statistic (`fun`) is computed for each resampled dataset,
and a histogram of the bootstrap sample statistics is plotted. A confidence interval is
calculated based on the `alpha` level and is displayed on the histogram along with the point estimate.
}
\examples{
set.seed(123)
x <- rnorm(30)
myboot2(iter=1000, x=x, fun="mean", alpha=0.05, cx=1.5)

}
