% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Calculate a 95\% Confidence Interval for the Mean}
\usage{
myci(x)
}
\arguments{
\item{x}{A numeric vector representing the sample data.}
}
\value{
A numeric vector of length 2, containing the lower and upper bounds of the 95\% confidence interval.
}
\description{
This function calculates a 95\% confidence interval for the population mean (\eqn{\mu})
based on a single sample. It assumes a normally distributed population.
}
\examples{
# Example usage with a sample of data
sample_data <- c(21.65, 17.48, 20.1, 21.57, 14.82)
myci(sample_data)
}
