% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Utilizes the Poisson distribution to normalize and sample and produce various graphs.}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{The sample input}

\item{iter}{Number of iterations}

\item{lambda}{Mean number of events occurring within a given interval}

\item{...}{Any other optional arguments}
}
\description{
Utilizes the Poisson distribution to normalize and sample and produce various graphs.
}
\examples{
mycltp(n=2, iter=10000, lambda = 4)

}
