% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Samples from range 1-n over k iterations, produces a graph for each iteration.}
\usage{
myf(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Sample size}

\item{iter}{Number of iterations}

\item{time}{Number of seconds between the creation of each plot}
}
\description{
Samples from range 1-n over k iterations, produces a graph for each iteration.
}
\examples{
myf(n=10,iter=20,time=1)

}
