% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Samples from range 1-n over k iterations, produces a graph for each iteration.}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean of the distribution}

\item{sigma}{standard deviation}

\item{a}{right most boundary of which probability is found under curve from -inf to a}
}
\description{
Samples from range 1-n over k iterations, produces a graph for each iteration.
}
\examples{
myncurve(mu=5,sigma=3,a=4)

}
