% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mypackage}
\alias{GeeStatsFall2024Package}
\alias{GeeStatsFall2024Package-package}
\alias{mypackage}
\title{MATH4753FALL2024}
\description{
Contains several functions from the OU MATH4753 course.
}
\author{
Dakota Gee \email{Dakota.L.Gee-1@ou.edu}
}
