% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Samples from range 1-n over k iterations, produces a graph for each iteration.}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats on the plane}

\item{gamma}{Maximum desired probability that plane will be overbooked}

\item{p}{Probability that someone shows up for their flight}
}
\description{
Samples from range 1-n over k iterations, produces a graph for each iteration.
}
\examples{
ntickets(N=200,gamma=0.02,p=0.95)

}
